Cambios a la vida del jugador
-Cambio en los scripts que recibían como valor un float a por int en los valores relacionados a la vida.
-Implementación del HUD y la vida del jugador en la parte superior izquierda de la pantalla del juego.

+++ Prog-Vj2/Assets/Script/Jugador/PerfilJugador.cs
@@ -44,8 +44,8 @@ public class PerfilJugador : ScriptableObject
     [Header("Configuracion de Atributos")]
     [SerializeField]
     [Range(5, 10)]
-    private float vida = 5f;
-    public float Vida { get => vida; set => vida = value; }
+    private int vida = 5;
+    public int Vida { get => vida; set => vida = value; }
 
     //Configuraciones de los Sonidos.
     [Header("Configuracion de Sonido")]
